refactor: split Dexie schema migration and seed setup out of db-setup.ts constructor (#623)
